+++ src/DataBlobs/DataBlobs.jl
@@ -20,4 +20,4 @@ export getData, addData!, updateData!, deleteData!, listStoreEntries
 export getDataBlob, addDataBlob!, updateDataBlob!, deleteDataBlob!, listDataBlobs
 export copyStore
 
-export getId, getHash, getCreatedTimestamp
+export getId, getHash, getTimestamp